Fix for int and double error and fix for mintemp in userplants

+++ client/lib/api/UserPlant.dart
@@ -8,7 +8,7 @@ class UserPlant implements JsonSerializeable {
   int id;
   int userId;
   String nickname;
-  int potVolume;
+  double potVolume;
   double latitude;
   double longitude;
   String imageName;
